Add alwaysFireChange option
Given I am using the picker for text, and my text has mutliple
colors (green, yellow). Then I am initializing the colorpicker with
the color that makes the majority of the text (green).

When I want to change the whole textcolor to green, which is
realized by the change-callback, it would not fire, because the
color green was choosen before as it was the major color.

This commit adds an option which will cause the event to fire
on every click.