Add keyboard navigation to the color palette, for accessibility.
Allow tabbing into the color picker, and focus to be set on a
square in the color picker.  Allow using arrow keys to move around
the color palette and set focus.  Allow Enter key to select the
focused color.  This is done with a new keydown handler.

This is needed in order to comply with accessibility requirements.
Note that the color gradient is not yet accessible, just the palette.

https://github.com/bgrins/spectrum/issues/295