Ability to change allowEmpty from false to true
There was a couple weird things going on that prevented this from happening in the current build. For one, if `allowEmpty` was false, `clearButton` was permanently hidden. After changing allowEmpty to true it was still empty, even with the `sp-clear-enabled` class applied. Two, allowEmpty is a global variable. Once it was false it was always false, so when trying to apply new options it's pretty much skipped over.