Remove stopPropagation on choose & cancel buttons
It's not ideal to use stopPropagation() on these two bottons in case if users of the library wishes to intercept the clicking event on them and perform explicit actions on colour picked through hex code, for example. There are also plenty of other places that use stopPropagation that may need some cleanup, but let's start with these two for now.